Changed clickTolerance behaviour
When clicking on a track with clickTolerance, features will now be returned if they fall within an area around the click whose width is the maximum of the feature's width in px, or the clickTolerance.
This allows tracks with a mixture of very small and larger features to have clickTolerance applied, without it affecting the larger features.