Fix broken Scaleline, Legend and Labels track tests
The tests were failing because of slight pixel offsets being incorrect.

It would probably be easier to debug these if we were generating and writing to snapshot files,
as we could then see the changes to be made to adjust the tests exactly.

As it stands, the tests probably contain a little too much logic, as they're reimplementing the drawing logic,
but for now this is the best way to fix the tests by adjusting the draw coordinates